+++ src/modules/fuzz/Fuzz.h
@@ -35,7 +35,7 @@ class Fuzz
         float getThreshold();
         void setThreshold( float value );
         
-        void apply( float* channelData, int bufferSize );
+        void apply( float* channelData, unsigned long bufferSize );
 
     private:
         float _amount;
